Move index.js to kappa-old.js
